requirements.txt
gigachain

python-docx